Tahmid Chowdhury With JavaScript Object
In JavaScript, an object is a collection of key-value pairs, often referred to as properties. The keys are strings (or Symbols), and the values can be of any data type, including other objects, functions (methods), or primitive values like numbers, strings, or booleans. Objects are fundamental for representing data structures and encapsulating related data and functionality.