Tahmid Chowdhury With  Breakdown_JavaScript
This text discusses various new features introduced in the ES6 version of JavaScript, including destructuring for arrays, functions, and objects, along with explanations on how to use it effectively. Additionally, it covers the rest and spread operators, detailing the rules for copying using the spread operator and for duplicating arrays with the rest operator. Furthermore, it addresses the rules for exporting and importing modules in JavaScript. A note is included that provides important information regarding import and export functionality.